resource opener reorganization
